+++ examples/init/routes/_app.tsx
@@ -8,6 +8,7 @@ export default function App({ Component }: PageProps) {
         <meta name='viewport' content='width=device-width, initial-scale=1.0' />
         <title>init</title>
         <link rel='stylesheet' href='/styles.css' />
+        <link rel='stylesheet' href='/fonts.css' />
       </head>
       <body>
         <Component />